+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Changelog
 * Add new argument ``legend`` to :meth:`pycov.CovFrame.plot_region` method.
 * Add new method :meth:`pyvcf.VcfFrame.fetch`.
 * Update :meth:`pyvcf.VcfFrame.strip` method to handle cases where one or more specified FORMAT keys are missing in a row.
+* Add new method :meth:`pyvcf.VcfFrame.pseudophase`.
 
 0.21.0 (2021-08-16)
 -------------------